expect this and get that

+++ 0x01-variables_if_else_while/1-last_digit.c
@@ -13,7 +13,7 @@ int main(void)
 	srand(time(0));
 	n = rand() - RAND_MAX / 2;
 	/* your code goes there */
-	long tmp = n % 10;
+	int tmp = n % 10;
 
 	if (tmp > 5)
 		printf("Last digit of %d is %d and is greater than 5\n", n, tmp);
